A sudden bug appeared where the screen was corrupted on cold start. If the power supply was cycled quickly after the corruption, it worked.
It looked like it was caused by slow or glitchy power rise. Added a big
delay before Main PIC sends data to LCD PIC, but no success. Left in place
as its not a bad to have a delay.

Changed initializeOutputs to write to latches instead of ports to avoid
read-modify-write problems caused by slow voltage shifts on the outputs.
This occurs when writing several pins in quick succession as each one
causes the entire port pins to be read and re-written -- if an output
setting has not caused the line to change (due to capacitance, wierd
voltages, etc.) then the read will read the unchanged line and then
write it back to the latch.

This seemed to solve the problem.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>