Updated method of displaying depth and target by creating functions to perform those tasks.
Updated incDepth and decDepth to call incDepthAbs and decDepthAbs.

Renamed incBCDAbs and decBCDAbs to incDepthABS and decDepthABS as depth
is the only variable that needs modifying. Functions still need to be
updated to use addBCDVars and subtractBCDVars.

Changed displayBCDVar to use FSR1 as a pointer for more efficient code.
Checked to verify that all functions called use FSR0, so no need to
save and restore FSR1.

Signed-off-by: Mike Schoonover <mike@skoonie.com>