Changed references to "depth" to "target". The code now uses "target" to refer to the user entered target depth and "depth" represents the actual position (or depth) of the cutting blade.
Changed setDepth function to setTarget. Changed that function to use
FSR1 to track the digit being modified instead of using scratch7. Function
adjustBCDDigit also changed to use FSR1.

Renamed isDepthZero to isTargetZero and added the code. It appears that
the old isDepthZero was an empty stub function with no code.

Added code to skip any software delays when "debug" is defined so that
simulating won't bog down in wait loops.

Signed-off-by: Mike Schoonover <mike@skoonie.com>