Changed code throughout the program to not rely on FSR0H being clear
Because the code was originally written for an older chip, when Mike S
went to use the code on the new chip, he just cleared the highs of FSR0
and FSR1 in start:.

This worked until we began using FSRn for other places in memory that
required the high bytes to be changed; the high bytes were no longer
cleared. To solve the problem and make the code more robust, everywhere
that uses FSRn as a pointer has been updated to set the high and low bytes.

Also todo (wip hss) comments.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>