Renamed debug to DEBUG_MODE to make it easier to search for.
Added a flag to signal when debounce timeout is counting down to improve
efficiency.

Changed decrement method of the debounce counter to subtract by 1
instead of decf which doesn't work well with multiple bytes.

Adjusted values used for debounce1:0 to reflect that the new code will
count another 255 bytes (as it should).

Signed-off-by: Mike Schoonover <mike@skoonie.com>