Fixed bug in setTarget. Call to processIO was changing FSR1...added code to push/pop FSR1.
Blocked code which was reading Select button...this button is now used to
read the AC OK signal from the Cutting Current power supply, so was being
in active state by R84 which is installed to handle the new input. The
Select button input now comes from the remote device such as a User
Interface board.

Blocked out call to sendDataIfReady because the LCD PIC was going crazy
(restarting?) when this was in place. The packets aren't even addressed
to the LCD PIC, so it must be responding in error.

Signed-off-by: Mike Schoonover <mike@skoonie.com>