Function jogMode now clears the DEBOUNCE_ACTIVE flag to disable debounce delays to improve responsiveness to the Up/Down switch.
Added delay at the beginning of jogMode to allow the user time to release
the Select button so mode doesn't exit immediately. The function disables
switch debouncing to increase responsiveness but this causes it to see
the Select switch input too long.

Add msDelay function. Need to remove other delay functions and use this
one instead as it is cleaner and simpler.

Signed-off-by: Mike Schoonover <mike@skoonie.com>